Move NewWalletPublicKeyHash to coordinator.go
