+++ pkg/coordinator/coordinator.go
@@ -1,7 +1,29 @@
 package coordinator
 
 import (
+	"fmt"
+
 	"github.com/ipfs/go-log"
+
+	"github.com/keep-network/keep-core/pkg/internal/hexutils"
 )
 
 var logger = log.Logger("keep-coordinator")
+
+// NewWalletPublicKeyHash decodes a wallet public key hash from a string.
+func NewWalletPublicKeyHash(str string) ([20]byte, error) {
+	var result [20]byte
+
+	walletHex, err := hexutils.Decode(str)
+	if err != nil {
+		return result, err
+	}
+
+	if len(walletHex) != 20 {
+		return result, fmt.Errorf("invalid bytes length: [%d], expected: [%d]", len(walletHex), 20)
+	}
+
+	copy(result[:], walletHex)
+
+	return result, nil
+}
